+++ packages/chains/bitcoin/lib/BitcoinChain.ts
@@ -341,10 +341,12 @@ class BitcoinChain extends AbstractUtxoChain<BitcoinTx, BitcoinUtxo> {
    * verifies additional conditions for a BitcoinTransaction
    * - check change box
    * @param transaction the PaymentTransaction
+   * @param signingStatus the signing status of transaction
    * @returns true if the transaction is verified
    */
   verifyTransactionExtraConditions = (
-    transaction: PaymentTransaction
+    transaction: PaymentTransaction,
+    signingStatus: SigningStatus = SigningStatus.UnSigned
   ): boolean => {
     const tx = Serializer.deserialize(transaction.txBytes);