+++ packages/chains/evm/lib/EvmChain.ts
@@ -724,12 +724,22 @@ abstract class EvmChain extends AbstractChain<Transaction> {
    * - multiple payments are not allowed in one transaction
    * - in case of erc-20 transfer, `data` must be appropriately parsed with the `transfer` ABI
    * @param transaction the PaymentTransaction
+   * @param signingStatus the signing status of transaction
    * @returns true if the transaction is verified
    */
   verifyTransactionExtraConditions = (
-    transaction: PaymentTransaction
+    transaction: PaymentTransaction,
+    signingStatus: SigningStatus = SigningStatus.UnSigned
   ): boolean => {
-    const tx = Serializer.deserialize(transaction.txBytes);
+    let tx: Transaction;
+    try {
+      tx = Serializer.deserialize(transaction.txBytes);
+    } catch (error) {
+      this.logger.warn(
+        `Tx [${transaction.txId}] is not verified: failed to deserialized due to error: ${error}`
+      );
+      return false;
+    }
 
     if (tx.to === null) {
       this.logger.warn(