+++ packages/chains/cardano/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rosen-chains/cardano
 
+## 10.1.0
+
+### Minor Changes
+
+- Add signingStatus to verifyTransactionExtraConditions function arguments
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-chains/abstract-chain@11.0.0
+
 ## 10.0.0
 
 ### Major Changes