+++ packages/chains/ethereum/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rosen-chains/ethereum
 
+## 0.2.0
+
+### Minor Changes
+
+- Add signingStatus to verifyTransactionExtraConditions function arguments
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-chains/abstract-chain@11.0.0
+  - @rosen-chains/evm@5.1.0
+
 ## 0.1.10
 
 ### Patch Changes